requirements.txt
# pip install -r requirements.txt

tox
pytest
aiohttp
python-mpv
pickle5
requests
RPi.GPIO
adafruit-blinka
adafruit-circuitpython-rgb-display
pillow